+++ src/org/wmbus/simulation/convergence/model/config/InitialSkipTimesAverageConvergenceConfigModel.java
@@ -1,14 +1,14 @@
 package org.wmbus.simulation.convergence.model.config;
 
 public class InitialSkipTimesAverageConvergenceConfigModel extends TimesAverageConvergenceConfigModel {
-    private int initialSkip = 0;
+    private long initialSkip = 0;
 
-    public InitialSkipTimesAverageConvergenceConfigModel(int initialSkip, double percentageConvergence, double percentageNotConvergence, int convergenceTimes, int notConvergenceTimes) {
+    public InitialSkipTimesAverageConvergenceConfigModel(long initialSkip, double percentageConvergence, double percentageNotConvergence, int convergenceTimes, int notConvergenceTimes) {
         super(percentageConvergence, percentageNotConvergence, convergenceTimes, notConvergenceTimes);
         this.initialSkip = initialSkip;
     }
 
-    public int getInitialSkip() {
+    public long getInitialSkip() {
         return initialSkip;
     }
